+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 unreleased
 ----------
+- Add `changeActiveView` event to track when the Drop-in updates the active view presented to the customer
 - Google Pay
   - Fix issue where passing custom button option would caused a developer error in the console (#701)